backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
pydantic[email]==2.10.3
pydantic-settings==2.7.0
python-dotenv==1.0.1
python-multipart==0.0.18
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.28.1
anthropic==0.42.0
pandas==2.2.3
openpyxl==3.1.5
pdfplumber==0.11.4
stripe>=11.0.0
